services/model-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
psutil==5.9.6
GPUtil==1.4.0
numpy==1.25.2
llama-cpp-python==0.2.20
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1